fixed punctuation splitting; separated auth/thread logic
